requirements.txt
django
django_cas_ng
django-cors-headers
djangorestframework
uwsgi
psycopg2
